refactor(game): move game files into own folder
