+++ src/components/game/board.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
 
-import '../index.scss';
+import '../../index.scss';
 import Square from './square';
 import { IBoardState, IPossibleMoves } from './game';
-import { coordinate, IPiece } from './pieces/IPieces.model';
+import { coordinate, IPiece } from '../pieces/IPieces.model';
 
 interface ISquare {
   style: string;
