+++ src/components/game/game.model.ts
