+++ src/components/game/home.tsx
